+++ central/main.go
@@ -67,7 +67,6 @@ import (
 	"github.com/stackrox/rox/central/pruning"
 	rbacService "github.com/stackrox/rox/central/rbac/service"
 	"github.com/stackrox/rox/central/reprocessor"
-	riskService "github.com/stackrox/rox/central/risk/service"
 	"github.com/stackrox/rox/central/role"
 	roleDataStore "github.com/stackrox/rox/central/role/datastore"
 	"github.com/stackrox/rox/central/role/mapper"
@@ -285,7 +284,6 @@ func (f defaultFactory) ServicesToRegister(registry authproviders.Registry) []pk
 		policyService.Singleton(),
 		processIndicatorService.Singleton(),
 		processWhitelistService.Singleton(),
-		riskService.Singleton(),
 		roleService.Singleton(),
 		rbacService.Singleton(),
 		sacService.Singleton(),