+++ ui/src/Containers/Risk/RiskSidePanel.js
@@ -1,8 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 
-import { fetchDeployment } from 'services/DeploymentsService';
-import fetchRisk from 'services/RisksService';
+import { fetchDeploymentWithRisk } from 'services/DeploymentsService';
 import { fetchProcesses } from 'services/ProcessesService';
 
 import Panel from 'Components/Panel';
@@ -11,7 +10,6 @@ import RiskSidePanelContent from './RiskSidePanelContent';
 function RiskSidePanel({ selectedDeploymentId, setSelectedDeploymentId }) {
     const [selectedDeployment, setSelectedDeployment] = useState(undefined);
     const [selectedProcesses, setSelectedProcesses] = useState(undefined);
-    const [selectedRisk, setSelectedRisk] = useState(undefined);
 
     const [isFetching, setIsFetching] = useState(false);
 
@@ -24,31 +22,22 @@ function RiskSidePanel({ selectedDeploymentId, setSelectedDeploymentId }) {
 
             setIsFetching(true);
             Promise.all([
-                fetchDeployment(selectedDeploymentId),
-                fetchRisk(selectedDeploymentId, 'deployment'),
+                fetchDeploymentWithRisk(selectedDeploymentId),
                 fetchProcesses(selectedDeploymentId)
             ]).then(
-                ([deployment, risk, processes]) => {
-                    setSelectedDeployment(deployment);
-                    setSelectedRisk(risk);
+                ([deploymentWithRisk, processes]) => {
+                    setSelectedDeployment(deploymentWithRisk);
                     setSelectedProcesses(processes.response);
                     setIsFetching(false);
                 },
                 () => {
                     setSelectedDeployment(undefined);
-                    setSelectedRisk(undefined);
                     setSelectedProcesses(undefined);
                     setIsFetching(false);
                 }
             );
         },
-        [
-            selectedDeploymentId,
-            setSelectedDeployment,
-            setSelectedRisk,
-            setSelectedProcesses,
-            setIsFetching
-        ]
+        [selectedDeploymentId, setSelectedDeployment, setSelectedProcesses, setIsFetching]
     );
 
     function unselectDeployment() {
@@ -57,17 +46,16 @@ function RiskSidePanel({ selectedDeploymentId, setSelectedDeploymentId }) {
 
     // Only render if we have image data to render.
     if (!selectedDeploymentId) return null;
-
     return (
         <Panel
-            header={!selectedDeployment ? 'Unknown Deployment' : selectedDeployment.name}
+            header={!selectedDeployment ? 'Unknown Deployment' : selectedDeployment.deployment.name}
             className="bg-primary-200 w-full h-full absolute pin-r pin-t md:w-1/2 min-w-72 md:relative"
             onClose={unselectDeployment}
         >
             <RiskSidePanelContent
                 isFetching={isFetching}
-                selectedDeployment={selectedDeployment}
-                deploymentRisk={selectedRisk}
+                selectedDeployment={!selectedDeployment ? null : selectedDeployment.deployment}
+                deploymentRisk={!selectedDeployment ? null : selectedDeployment.risk}
                 processGroup={selectedProcesses}
             />
         </Panel>