+++ ui/src/reducers/index.js
@@ -13,7 +13,6 @@ import notifications, { selectors as notificationSelectors } from './notificatio
 import featureFlags, { selectors as featureFlagSelectors } from './featureFlags';
 import globalSearch, { selectors as globalSearchSelectors } from './globalSearch';
 import policies, { selectors as policySelectors } from './policies/reducer';
-import risks, { selectors as riskSelectors } from './risks';
 import roles, { selectors as roleSelectors } from './roles';
 import searchAutoComplete, { selectors as searchAutoCompleteSelectors } from './searchAutocomplete';
 import serverError, { selectors as serverErrorSelectors } from './serverError';
@@ -46,7 +45,6 @@ const appReducer = combineReducers({
     globalSearch,
     cli,
     policies,
-    risks,
     roles,
     searchAutoComplete,
     serverError,
@@ -86,7 +84,6 @@ const getNotifications = state => getApp(state).notifications;
 const getFeatureFlags = state => getApp(state).featureFlags;
 const getGlobalSearches = state => getApp(state).globalSearch;
 const getPolicies = state => getApp(state).policies;
-const getRisks = state => getApp(state).risks;
 const getRoles = state => getApp(state).roles;
 const getSearchAutocomplete = state => getApp(state).searchAutoComplete;
 const getServerError = state => getApp(state).serverError;
@@ -115,7 +112,6 @@ const boundSelectors = {
     ...bindSelectors(getFeatureFlags, featureFlagSelectors),
     ...bindSelectors(getGlobalSearches, globalSearchSelectors),
     ...bindSelectors(getPolicies, policySelectors),
-    ...bindSelectors(getRisks, riskSelectors),
     ...bindSelectors(getRoles, roleSelectors),
     ...bindSelectors(getRoute, routeSelectors),
     ...bindSelectors(getSearchAutocomplete, searchAutoCompleteSelectors),